// apps/api/src/users/user.resource.ts
// Auto-generated by json-api-codegen
// Do not edit manually

import type { UserEntity } from './user.entity';

export interface UserAttributes {
  name: string;
  email: string;
  role: 'admin' | 'user';
}

export interface UserResource {
  id: string;
  type: 'user';
  attributes: UserAttributes;
}

export const asUserResource = (user: UserEntity): UserResource => ({
  id: user.id,
  type: 'user',
  attributes: {
    name: user.name,
    email: user.email,
    role: user.role,
  },
});